Convert Date
In converters.js we imported moment library and set set DateFormatValueConverter to return only month date and year. Aurelia can recognize any class that ends with ValueConverter. This is why our class name is DateFormatValueConverter. This class will be registered as dateFormat and we can later use it inside view